Add files via upload
The evaluation results of homicide in three significant Canadian provinces during the course of the 
previous 10 years, from 2011 to 2021, are presented in this paper. The federal police of Canada continue 
to prioritize internal security, and this research examines several facets of the rise in homicide cases in 
the three provinces that were the focus of this analysis, namely Ontario, Alberta, and Manitoba. The 
number of offenses recorded on the Stats Canada website served as the methodology for this research. 
The review focused on four main issues: crime rates in the subject provinces for the previous ten years, 
the province’s most violent cities, any relationships between juvenile and adult homicide cases, and 
projections of crime rates for the following year.